+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 [unreleased]
 
 - Implemented `UpdateEl` for `Filter` and `FilterMap`.
+- Added method `El::is_custom(&self)`.
+- Fixed custom elements patching (#325).
 
 ## v0.5.1
 - [BREAKING] `MessageMapper::map_message` changed to `MessageMapper::map_msg`.